AI-Bootcamp-Demo/requirements.txt
h5py==2.8.0
html5lib==1.0.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
matplotlib==2.2.2
numpy==1.14.5
opencv-python==3.4.2.17
pandas==0.23.3
Pillow==5.2.0
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
sklearn==0.0
tensorflow==1.9.0
urllib3==1.23
gunicorn==19.7
flask==0.12.2
flask-cors==3.0.3
azureml-sdk
